deleted unnesesary files norm check few funcrions separated in 2

+++ srcs/parser/parser_clear.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
-/*                                                        ::::::::            */
-/*   parser_clear.c                                     :+:    :+:            */
-/*                                                     +:+                    */
-/*   By: znajda <znajda@student.codam.nl>             +#+                     */
-/*                                                   +#+                      */
-/*   Created: 2022/08/31 13:29:59 by znajda        #+#    #+#                 */
-/*   Updated: 2022/09/17 16:00:24 by znajda        ########   odam.nl         */
+/*                                                        :::      ::::::::   */
+/*   parser_clear.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: yuliia <yuliia@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/08/31 13:29:59 by znajda            #+#    #+#             */
+/*   Updated: 2022/09/18 11:56:40 by yuliia           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,16 +15,12 @@
 #include <stdlib.h>
 #include "structs.h"
 #include "utils.h"
-#include <parser_all.h>
-#include <lexer_all.h>
+#include "parser_all.h"
+#include "lexer_all.h"
 
-t_parse	*rm_one_from_parse_list(t_parse *to_remove)
+//split rm_one_from_parse_list function
+void	free_cmd_args_outfile_infile(t_parse *to_remove)
 {
-	t_parse	*to_return;
-
-	if (!to_remove)
-		return (NULL);
-	to_return = to_remove->next;
 	if (to_remove->cmd)
 		free(to_remove->cmd);
 	to_remove->cmd = NULL;
@@ -37,6 +33,16 @@ t_parse	*rm_one_from_parse_list(t_parse *to_remove)
 	if (to_remove->infile)
 		free(to_remove->infile);
 	to_remove->infile = NULL;
+}
+
+t_parse	*rm_one_from_parse_list(t_parse *to_remove)
+{
+	t_parse	*to_return;
+
+	if (!to_remove)
+		return (NULL);
+	to_return = to_remove->next;
+	free_cmd_args_outfile_infile(to_remove);
 	if (to_remove->heredoc)
 	{
 		unlink(to_remove->heredoc);
